\name{show.year.month.season}
\alias{show.year.month.season}
\title{Show 'year','month','season'}
\usage{
show.year.month.season(data, TimeField = "Time", add_fields = c("month",
  "year", "season"))
}
\arguments{
  \item{data}{data frame}

  \item{TimeField}{character string containg time field of
  \code{data}}

  \item{add_fields}{character string. Default is
  \code{c("month","year","season")}}
}
\description{
Add a column indicating year,month or season
}

